Standardise space, pop the helper class out
